feat: created ResourceTexture class
the cpp file is empty because I want to discuss how we want to set the members (width, format, etc.). Do we add them to the constructor, meaning they'll be computed in the Module, or do we compute them inside the ResourceTexture class? In the first case the cpp won't be needed, but in the second one it will