fix [lights]: "fixed" spotlight angle
Capped the maximum angle for the spotlights to 90 degrees. There wasn't actually anything to fix in this case, when the angle for the spotlights was greater than 90 degrees (1.57 radians), the effect that seemed to bug out the light was actually the light bending into itself (?) 'cause the angle was greater than 90 degrees, causing the light to shadow itself (like the effect of a black whole curving light around it in its event horizon).

Other than capping the light, i've changed the imgui to show the values in degrees instead of radians, so it's more clear the angle that is being used for the light.